+++ src/functions/google-pubsub/index.ts
@@ -1,0 +1,1 @@
+export { GooglePubSubWriteTopics } from './write-topics.js';
